requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
groq==0.11.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.4
Werkzeug==3.0.1